Phase 3B Split 10: extract EngineFxBuilder from GhostVisualBuilder
TryBuildEngineFX (565 lines) + 5 helper methods moved to new
EngineFxBuilder.cs static class. Local function closures stay
as local functions (they capture loop variables, not class state).
Also moved: fxModelRotationFallbackEuler dict, EngineModelNodeTypes
array, TryGetFxModelFallbackRotation method.

8 shared helpers promoted from private to internal in
GhostVisualBuilder: FindFxPrefab, NormalizeFxPrefabName,
StripKspFxControllers, ConfigureGhostEngineParticleSystems,
ResolveGhostFxParent, LogFxParentAlignmentDiagnostic,
LogFxInstancePlacementDiagnostic, FindTransformsRecursive.

GhostVisualBuilder: 7600 -> 6625 lines (-975).

Build: 0 errors, 0 warnings. Tests: 3261/3261 pass.
